requirements.txt
pdf2image==1.17.0
pdfminer.six==20240706
openai>=1.35.7
scikit-learn>=1.3.2
rich>=13.7.1
tqdm>=4.66.4
pandas>=2.2.2
python-dotenv>=1.0.1
Pillow>=10.3.0
numpy>=1.26.4
